conformance/functionalityTesting/basics/createContext.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>createContext conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("CreateContext functionality test case.");

    var wtu = WebCLTestUtils;
    var webCLPlatform = wtu.getPlatform();
    var webCLContext, contextDevices;

    var createContext = function(param1, param2) {
        try {
            webCLContext = wtu.createContext(param1, param2);
            contextDevices = webCLContext.getInfo(webcl.CONTEXT_DEVICES);

            if (param1 == undefined) {
                var defaultDevice = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_DEFAULT)[0];
                if (contextDevices[0].getInfo(webcl.DEVICE_TYPE) !== defaultDevice.getInfo(webcl.DEVICE_TYPE)) {
                    console.log("Failed in validating the number of devices, when no argument is passed.");
                    return null;
                }
            } else if (typeof(param1) == 'number') {
                if (param1 == webcl.DEVICE_TYPE_DEFAULT)
                    return webCLContext;
                if (contextDevices[0].getInfo(webcl.DEVICE_TYPE) !== param1) {
                    console.log("Falied in validating createContext on devieType. Expected: " + param1 + ". Obtained : " + contextDevices[0].getInfo(webcl.DEVICE_TYPE));
                    return null;
                }
            } else if (param1 instanceof Array) {
                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i] !== param1[i]) {
                        console.log("Failed in validating devices[" + i + "]. Expected : " + param1[i] + ". Obtained : " + contextDevices[i]);
                        return null;
                    }
            } else if (Object.getPrototypeOf(param1).toString() == '[object WebCLDevicePrototype]') {
                if (contextDevices[0] !== param1) {
                    console.log("Failed in validating devices[" + i + "]. Expected : " + param1[i] + ". Obtained : " + contextDevices[i]);
                    return null;
                }
            } else if (Object.getPrototypeOf(param1).toString() == '[object WebCLPlatformPrototype]') {
                if (contextDevices[0].getInfo(webcl.DEVICE_PLATFORM) === param1) {
                    if (param2) {
                        if (contextDevices[0].getInfo(webcl.DEVICE_TYPE) !== param2) {
                            console.log("Failed in validating deviceType used to create WebCLContext. Expected : " + param2 + ". Obtained : " + contextDevices[0].getInfo(webcl.DEVICE_TYPE));
                            return null;
                        }
                    }
                } else {
                    console.log("Failed in validating platform used to create WebCLContext.  Expected : " + param1 + ". Obtained : " + contextDevices[i].getInfo(webcl.DEVICE_PLATFORM));
                    return null;
                }
            }
            return webCLContext;
        } catch(e) {
            e.description = "Exception thrown while creating and validating context : " + e.name;
            throw e;
        }
    }

    shouldBeType("createContext()", "WebCLContext");
    shouldBeType("createContext(webCLPlatform)", "WebCLContext");

    /*
     *  Creating context on type : CPU
     */
    shouldBeTypeOrException("createContext(webcl.DEVICE_TYPE_CPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(webCLPlatform, webcl.DEVICE_TYPE_CPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_CPU)[0])", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_CPU))", "WebCLContext", "DEVICE_NOT_FOUND");

    /*
     *  Creating context on type : GPU
     */
    shouldBeTypeOrException("createContext(webcl.DEVICE_TYPE_GPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(webCLPlatform, webcl.DEVICE_TYPE_GPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_GPU)[0])", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_GPU))", "WebCLContext", "DEVICE_NOT_FOUND");

    /*
     *  Creating context on type : ACCELERATOR
     */
    shouldBeTypeOrException("createContext(webcl.DEVICE_TYPE_ACCELERATOR)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(webCLPlatform, webcl.DEVICE_TYPE_ACCELERATOR)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ACCELERATOR)[0])", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ACCELERATOR))", "WebCLContext", "DEVICE_NOT_FOUND");

    /*
     *  Creating context on type : DEFAULT
     */
    var webCLDevicesDEFAULT = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_DEFAULT);
    shouldBeType("createContext(webcl.DEVICE_TYPE_DEFAULT)", "WebCLContext");
    shouldBeType("createContext(webCLDevicesDEFAULT)", "WebCLContext");

    /*
     *  Creating context on type : ALL
     */
    var webCLDevicesALL = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ALL);
    shouldBeType("createContext(webCLDevicesALL)", "WebCLContext");

    /*
     *  Creating context with at least one available valid device type.
     */

    var flag = false, webCLContextOnDeviceType;
    var deviceTypeArray = [webcl.DEVICE_TYPE_GPU, webcl.DEVICE_TYPE_CPU, webcl.DEVICE_TYPE_ACCELERATOR];
    for (var i = 0; i < deviceTypeArray.length; i++) {
        try {
            webCLContextOnDeviceType = webcl.createContext(deviceTypeArray[i]);
        } catch(e) {
            console.log("Exception thrown for device type : " + deviceTypeArray[i] + " : "  + e.name);
        }
        if (webCLContextOnDeviceType instanceof WebCLContext) {
            flag = true;
            testPassed("createContext for " + deviceTypeArray[i] + ".");
            break;
        }
    }
    if (!flag)
        testFailed("Failed to create context on any of device {GPU, CPU, ACCELERATOR};");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
